Backend/app/requirements.txt
fastapi==0.79.1
uvicorn==0.18.2
python-dotenv==0.20.0
scryp
pydantic